Instance.HttpScore: float64

+++ client.go
@@ -29,7 +29,7 @@ type Instance struct {
 	Name              string  `json:"name"`
 	Uptime            float64 `json:"uptime"`
 	Up                bool    `json:"up"`
-	HttpsScore        int     `json:"https_score"`
+	HttpsScore        float64 `json:"https_score"`
 	HttpsRank         string  `json:"https_rank"`
 	IPv6              bool    `json:"ipv6"`
 	OpenRegistrations bool    `json:"openRegistrations"`
